move tests to SelectSqlGenerationTest
